+++ packages/modulisto_flutter/lib/src/listenable_adapter.dart
@@ -1,35 +1,31 @@
 import 'package:flutter/foundation.dart';
-import 'package:modulisto/modulisto.dart';
+import 'package:modulisto/modulisto.dart' as modulisto;
 
-class ListenableAdapter implements Listenable, UnitHost {
+class ListenableAdapter implements Listenable {
   ListenableAdapter(this._unit);
 
-  final Notifiable<Object?> _unit;
+  final modulisto.ValueListenable<Object?> _unit;
   late final _realListeners = <VoidCallback>[];
-
-  @override
-  @protected
-  late final Map<Notifiable<Object?>, List<void Function(Object? p1)>> listeners = {};
 
   @override
   void addListener(VoidCallback listener) {
     if (_realListeners.isEmpty) {
-      _unit.notifiableHosts.add(this);
-      listeners.putIfAbsent(_unit, () => []).add((_) {
-        for (final callback in _realListeners) {
-          callback();
-        }
-      });
+      _unit.addListener(_callback);
     }
     _realListeners.add(listener);
+  }
+
+  void _callback(_) {
+    for (final callback in _realListeners) {
+      callback();
+    }
   }
 
   @override
   void removeListener(VoidCallback listener) {
     _realListeners.remove(listener);
     if (_realListeners.isEmpty) {
-      _unit.notifiableHosts.remove(this);
-      listeners.clear();
+      _unit.removeListener(_callback);
     }
   }
 }